requirements.txt
numpy==1.23.5
Pillow==9.3.0
scipy==1.9.3
